feat: Add AnalyzeJavaScript utility to Utils
This commit adds a new utility function called AnalyzeJavaScript to the Utils module. The function is exported from the index.ts file in the Utils directory.

The purpose of this change is to provide a utility function that can analyze JavaScript code. This function will be used in other parts of the codebase to perform static analysis on JavaScript files.

This commit also includes updates to the npmignore file to exclude the "exemple" directory from the npm package.

Lastly, there are some new files added to the project, including StartNode.tsx in the Components/Nodes directory, AddButton.tsx in the Components directory, and index.tsx in the Components directory.

These changes were made based on the recent user commit "first commit" and the established commit message conventions in the repository.