feat: Add vertical layout option for ReactFlowUI
This commit adds the option to set the layout of ReactFlowUI to "vertical" in the App component. This allows the UI to be displayed vertically instead of the default horizontal layout.

The purpose of this change is to provide more flexibility in how the UI is presented to users. Some users may prefer a vertical layout for better readability or to fit their specific use case.

These changes were made based on the recent user commit "feat: Add AnalyzeJavaScript utility to Utils" and the established commit message conventions in the repository.