feat: Update NodeField to include hidden property
This commit adds the hidden property to the NodeField component in the lib directory. The hidden property allows for fields to be hidden from view when rendering the component.

The purpose of this change is to provide more control over the visibility of fields in the NodeField component. By setting the hidden property to true, fields can be hidden from the user interface.

These changes were made based on the recent user commit "feat: Update collapse behavior in ActionNode and ConditionNode components" and the established commit message conventions in the repository.